# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MySchool API",
    "version": "2.2.0",
    "description": "API REST complète pour MySchool e-learning. \n        Gestion des utilisateurs, cours, chapitres, leçons, exercices, inscriptions et instructeurs.\n        Support complet des requêtes filtrées et index Firestore optimisés.",
    "contact": {
      "name": "MySchool Support",
      "email": "support@myschool.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-myschool-f862b.cloudfunctions.net",
      "description": "Serveur de production (Firebase Functions)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "phone",
          "level",
          "role",
          "status",
          "classe"
        ],
        "properties": {
          "uid": {
            "type": "string",
            "description": "ID Firebase Auth (auto-généré)"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "+221771234567"
          },
          "login": {
            "type": "string",
            "example": "+221771234567"
          },
          "password": {
            "type": "string",
            "example": "Password123",
            "description": "Requis à la création uniquement"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "beginner"
          },
          "classe": {
            "type": "string",
            "example": "3ème (BFEM)",
            "description": "Classe précise de l'utilisateur (obligatoire)"
          },
          "niveauScolaire": {
            "type": "string",
            "enum": [
              "ELEMENTAIRE",
              "MOYEN",
              "SECONDAIRE",
              "UNIVERSITAIRE"
            ],
            "example": "MOYEN"
          },
          "role": {
            "type": "object",
            "properties": {
              "libelle": {
                "type": "string",
                "enum": [
                  "student",
                  "instructor",
                  "admin"
                ],
                "example": "student"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ],
            "example": "active"
          },
          "profileImage": {
            "type": "string",
            "description": "URL de l'image de profil"
          },
          "specializationId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Referral": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID du parrainage"
          },
          "referrerId": {
            "type": "string",
            "description": "ID de l'utilisateur parrain"
          },
          "referredUserId": {
            "type": "string",
            "description": "ID de l'utilisateur parrainé"
          },
          "referralCode": {
            "type": "string",
            "example": "REF_GPL8_A1B2C3",
            "description": "Code de parrainage (REF_XXXX_XXXXXX)"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "COMPLETED",
              "EXPIRED",
              "CANCELLED"
            ],
            "example": "PENDING",
            "description": "Statut du parrainage"
          },
          "bonusAmount": {
            "type": "number",
            "example": 500,
            "description": "Montant du bonus en centimes"
          },
          "bonusType": {
            "type": "string",
            "enum": [
              "DISCOUNT",
              "FREE_COURSE",
              "POINTS"
            ],
            "example": "DISCOUNT",
            "description": "Type de bonus"
          },
          "clicks": {
            "type": "number",
            "example": 0,
            "description": "Nombre de clics sur le lien"
          },
          "conversions": {
            "type": "number",
            "example": 0,
            "description": "Nombre de conversions réussies"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "deepLink": {
                "type": "string",
                "example": "myschool://referral?code=REF_GPL8_A1B2C3"
              },
              "webLink": {
                "type": "string",
                "example": "https://myschool-app.com/join?ref=REF_GPL8_A1B2C3"
              },
              "shareText": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "example": "whatsapp"
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Message d'erreur"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/enrollments/user/{userId}": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "summary": "Récupérer les inscriptions d'un utilisateur",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des inscriptions"
          }
        }
      }
    },
...
    "/api/payments/user/{userId}": {
      "get": {
        "summary": "Récupère tous les paiements d'un utilisateur",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        }
      }
    },
...
    "/push-notifications/user/{userId}": {
      "get": {
        "summary": "Récupérer les notifications d'un utilisateur",
        "tags": [
          "Push Notifications"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des notifications"
          }
        }
      }
    },
...
    "/api/referrals/user/{userId}": {
      "get": {
        "summary": "Récupère tous les parrainages d'un utilisateur",
        "tags": [
          "Referrals"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'utilisateur"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des parrainages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Referral"
                      }
                    },
                    "count": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
...
    "/api/subscriptions/user/{userId}": {
      "get": {
        "summary": "Obtenir tous les abonnements d'un utilisateur",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'utilisateur"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des abonnements"
          }
        }
      }
    },
...
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Créer un nouvel utilisateur",
        "description": "Crée un utilisateur avec authentification Firebase automatique",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer tous les utilisateurs",
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "count": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/classes-disponibles/{niveau}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtenir les classes pour un niveau spécifique",
        "parameters": [
          {
            "in": "path",
            "name": "niveau",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ELEMENTAIRE",
                "MOYEN",
                "SECONDAIRE",
                "UNIVERSITAIRE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des classes pour ce niveau"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer un utilisateur par ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID du document Firestore"
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur trouvé"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Mettre à jour un utilisateur",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur mis à jour"
          },
          "400": {
            "description": "Erreur de validation"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Supprimer un utilisateur",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur supprimé"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/users/email/{email}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer un utilisateur par email",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur trouvé"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        }
      }
    },
    "/api/users/phone/{phone}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer un utilisateur par téléphone",
        "parameters": [
          {
            "in": "path",
            "name": "phone",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "+221771234567"
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur trouvé"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        }
      }
    },
    "/api/users/role/{role}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer les utilisateurs par rôle",
        "parameters": [
          {
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "instructor",
                "admin"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs"
          }
        }
      }
    },
    "/api/users/status/{status}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer les utilisateurs par statut",
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "suspended"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs"
          }
        }
      }
    },
    "/api/users/{id}/profile-image": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Mettre à jour l'image de profil",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profileImageBase64": {
                    "type": "string",
                    "description": "Image en base64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image mise à jour"
          }
        }
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Mettre à jour le statut",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive",
                      "suspended"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut mis à jour"
          }
        }
      }
    },
...
  }
}